+++ packages/sdk/src/types/attributes/index.ts
@@ -2,3 +2,4 @@ export * from "./ConnectorAttribute";
 export * from "./ConnectorAttributes";
 export * from "./requests/CreateAttributeRequest";
 export * from "./requests/GetAttributesRequest";
+export * from "./requests/GetValidAttributesRequest";