+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.0.0-alpha.5
 
--   upgrade the runtime to version 2.0.0-alpha.14
+-   upgrade the runtime to version 2.0.0-alpha.15
 -   add attribute routes
 
 ## 3.0.0-alpha.4